production/requirements.txt
# Production API Requirements
# ===========================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ML Models
openai-whisper==20231117
transformers==4.36.0
torch>=2.0.0
sentencepiece==0.1.99

# Audio Processing
soundfile==0.12.1
scipy==1.11.4
numpy<2.0

# Redis (for caching and rate limiting)
redis==5.0.1

# Security
PyJWT==2.8.0
python-multipart==0.0.6

# Monitoring
psutil==5.9.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-mock==3.12.0
httpx==0.26.0

# Optional: File type detection
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'
